requirement.txt
Flask==3.0.0
scikit-learn==1.5.0
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4
seaborn==0.13.2
mysql-connector-python==9.0.0
joblib==1.4.2
gunicorn==22.0.0